backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pydantic[email]==2.12.5
httpx==0.28.1
python-dotenv==1.2.1
motor==3.3.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.21
PyPDF2==3.0.1
openpyxl==3.1.5
